requirements.txt
discord.py
python-dotenv
sqlite3
asyncio
